some changes

+++ pages/api/products/[id].js
@@ -1,11 +1,13 @@
-import Product from 'models/Product'
-import dbConnect from 'utils/mongo'
+import Product from '../../../models/Product'
+import dbConnect from '../../../util/mongo'
 
 export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req
+  const token = cookies.token
 
   dbConnect()
 
@@ -19,19 +21,26 @@ export default async function handler(req, res) {
   }
 
   if (method === 'PUT') {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json('Not authenticated!')
+    }
     try {
-      const product = await Product.findyByIdAndUpdate(id, req.body, {
+      const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
       })
-      res.status(201).json(product)
+      res.status(200).json(product)
     } catch (err) {
       res.status(500).json(err)
     }
   }
+
   if (method === 'DELETE') {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json('Not authenticated!')
+    }
     try {
       await Product.findByIdAndDelete(id)
-      res.status(200).json('The product has been Deleted!')
+      res.status(200).json('The product has been deleted!')
     } catch (err) {
       res.status(500).json(err)
     }
